Return all test statistics values

Return all test statistics values

Usage

all_tests(object, svy_design = NULL, sim = NULL)

Arguments

object

A lavaan::lavaan() fit object.

svy_design

A survey::svydesign() object. Optional.

sim

(integer) Optional and used for large-scale simulations.

Value

Additionally, if sim argument is provided, two columns are appended: Whether the lavaan::lavaan() fit has converged and the matrix rank of \(\Omega_2\) (useful to see if any computational issues arose during model fit.)

Examples

fit <- lavaan::sem(txt_mod(1), gen_data_bin(1, n = 500), std.lv = TRUE,
                   estimator = "PML")
all_tests(fit)
#> # A tibble: 7 × 6
#>       W    df name           pval Xi_rank     S
#>   <dbl> <dbl> <chr>         <dbl>   <int> <int>
#> 1  3.17  5    Wald          0.674      12    15
#> 2  1.39  3.32 WaldV2,MM3    0.761      15    15
#> 3  3.16  5    WaldV3        0.675       5    15
#> 4  3.13  4.15 Pearson       0.559      15    15
#> 5  1.92  2.74 PearsonV2,MM3 0.541      15    15
#> 6  2.51  3.66 RSS,MM3       0.588      15    15
#> 7  3.07  5.01 Multn,MM3     0.690      15    15