Function to return the textual model for lavaan fit, depending on which simulation scenario we are investigating.

Function to return the textual model for lavaan fit, depending on which simulation scenario we are investigating.

Usage

txt_mod(model.no = 1L)

Arguments

model.no

(integer) Choose from 1--5. See details.

Value

Character vector of factor model to input to lavaan::lavaan().

Examples

txt_mod(1)
#> [1] "eta1 =~ NA*y1 + y2 + y3 + y4 + y5"